Transfer payment method (#282)
