+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/ui/CustomSaveCardActivity.kt
@@ -1,9 +1,13 @@
 package com.verygoodsecurity.vgscheckout.ui
 
 import android.os.Bundle
+import com.verygoodsecurity.vgscheckout.R
 import com.verygoodsecurity.vgscheckout.analytic.event.InitEvent
 import com.verygoodsecurity.vgscheckout.config.VGSCheckoutCustomConfig
 import com.verygoodsecurity.vgscheckout.ui.core.BaseCheckoutActivity
+import com.verygoodsecurity.vgscheckout.ui.fragment.core.BaseFragment
+import com.verygoodsecurity.vgscheckout.ui.fragment.method.SelectPaymentMethodFragment
+import com.verygoodsecurity.vgscheckout.ui.fragment.save.SaveCardFragment
 
 internal class CustomSaveCardActivity : BaseCheckoutActivity<VGSCheckoutCustomConfig>() {
 
@@ -11,4 +15,19 @@ internal class CustomSaveCardActivity : BaseCheckoutActivity<VGSCheckoutCustomCo
         super.initView(savedInstanceState)
         config.analyticTracker.log(InitEvent(InitEvent.ConfigType.CUSTOM, config))
     }
+
+    override fun navigateToSaveCard() {
+        val fragment = BaseFragment.create<SaveCardFragment>(config)
+        supportFragmentManager.beginTransaction()
+            .replace(R.id.fcvContainer, fragment, FRAGMENT_TAG)
+            .commit()
+    }
+
+    override fun navigateToPaymentMethods() {
+        val fragment = BaseFragment.create<SelectPaymentMethodFragment>(config)
+        supportFragmentManager.beginTransaction()
+            .replace(R.id.fcvContainer, fragment, FRAGMENT_TAG)
+            .commit()
+    }
+
 }
