+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/ui/fragment/method/adapter/PaymentMethodsAdapter.kt
@@ -58,6 +58,8 @@ internal class PaymentMethodsAdapter constructor(
         notifyDataSetChanged()
     }
 
+    fun hasSelectedCard(): Boolean = cards.getOrNull(selectedPosition) == null
+
     fun getSelectedPosition() = selectedPosition
 
     fun getSelectedCard(): Card? = cards.getOrNull(selectedPosition)
