--- vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/ui/fragment/orchestration/payment/PaymentFragment.kt
@@ -1,8 +1,0 @@
-package com.verygoodsecurity.vgscheckout.ui.fragment.orchestration.payment
-
-import com.verygoodsecurity.vgscheckout.config.core.CheckoutConfig
-import com.verygoodsecurity.vgscheckout.ui.fragment.core.BaseFragment
-
-internal class PaymentFragment : BaseFragment<CheckoutConfig>() {
-    // TODO: implement UI
-}
